Resolved Issue #65 added support for rhs and coef updates in GLPK.
Amended the test case so that it doesn't skip any sections for GLPK.

Also make a slight change to the exportModel. Previous code was accessing
an internal API method (those that are prefixed with an _glp) used the expected
code path GLPK.glp_write_lp. Note API docs say the 2nd argument is a hook
for future features and should be null (even if it does look a bit unseemly).